Fast Track: address review — decouple routing type, slim CI entry
Move VerdictRouting into shared/verdict.ts so buildVerdict and
buildOptOutVerdict no longer couple through a sibling builder module.
Extract parsePrContext() in ci.ts so main() reads as pure orchestration
and stays under the 20-line guideline.
